Run ChIP-seq analysis

Complete pipeline from quality assessment of raw reads through to peak calling and differential analysis.

run_chip(sample.info, reference = NULL, species = c("human", "mouse"),
  output.dir, threads = NULL, bigwig = FALSE, fastqc = "fastqc",
  multiqc = "multiqc", fastp = "fastp", samtools = "samtools",
  hisat2 = "hisat2", sambamba = "sambamba", idx = NULL,
  macs2 = "macs2", featurecounts = "featureCounts", broad = FALSE)

Arguments

sample.info
character string giving the path to a tab-delimited text file with at least the columns <condition> (treatment condition), <sample> (sample name), and <file1> (absolute or relative path to the fastq files). If fastq files and PE reads, then a column <file2> should also be present. If a batch effect is to be included in the design, then this should be identified under the column <batch>. If IP controls are used then these should be specified in columns <input1> (and <input2> if paired reads).

reference
character vector specifying the conditions in order. For example, c("A", "B", "C", "D") would mean "A" is the reference condition to which "B", "C" and "D" are compared; in addition, "C" and "D" will be compared to "B", and "D" will be compared to "C". If NULL then the comparisons will be arranged alphabetically. [DEFAULT = NULL].

species
character string specifying the name of the species. Only 'human', and 'mouse' are supported at present. [DEFAULT = human].

output.dir
character string specifying the directory to which results will be saved. If the directory does not exist, it will be created.

threads
an integer value indicating the number of parallel threads to be used by FastQC. [DEFAULT = maximum number of available threads - 1].

bigwig
logical, if TRUE then bigwif files will be created.

fastqc
a character string specifying the path to the fastqc executable. [DEFAULT = "fastqc"].

multiqc
a character string specifying the path to the multiqc executable. [DEFAULT = "multiqc"].

fastp
a character string specifying the path to the fastp executable. [DEFAULT = "fastp"].

samtools
a character string specifying the path to the samtools executable. [DEFAULT = "samtools"].

hisat2
a character string specifying the path to the hisat2 executable. [DEFAULT = "hisat2"].

sambamba
a character string specifying the path to the sambamba executable. [DEFAULT = "sambamba"].

idx
character vector specifying the basename of the index for the reference genome. The basename is the name of any of the index files up to but not including the final .1.ht2, etc. If NULL then the index for the relevant species (human or mouse) will be created using the build_index() function.

macs2
a character string specifying the path to the MACS2 executable. [DEFAULT = "macs2"].

featurecounts
a character string specifying the path to the featurecounts executable. [DEFAULT = "featureCounts"].

broad
logical, if TRUE then broad peaks will be called.