Perform peakcalling with MACS2

Perform peakcalling with MACS2

run_macs2(macs2 = "macs2", out.dir = "peaks", treatment.files = NULL,
  control.files = NULL, input.format = "AUTO", broad = FALSE,
  species = c("human", "mouse"), sample.names = NULL,
  no.model = FALSE, extsize = 0, shift = 0, no.lambda = FALSE,
  call.summits = TRUE, qvalue = 0.05, verbosity = 3,
  threads = NULL)

Arguments

macs2
character string specifying path to macs2 [DEFAULT = "macs2"]

out.dir
character string specifying the directory to which results will be saved. If the directory does not exist, it will be created. [DEFAULT = "peaks"]

treatment.files
character vector specifying path to alignment files. [DEFAULT = NULL]

control.files
character vector specifying path to input controls. This is optional, but if used then MUST be same length as <treatment.files>, so if the same input control is used for multiple ltreatments, then it should be specified multiple times. If NULL, then no.lambda will be set to TRUE. [DEFAULT = NULL]

input.format
character vector specifying format of input file; can be "ELAND", "BED", "ELANDMULTI", "ELANDEXPORT", "ELANDMULTIPET" (for pair-end tags), "SAM", "BAM", "BOWTIE", "BAMPE" or "BEDPE". [DEFAULT = "AUTO"]

broad
logical, if TRUE then MACS2 will output broadpeak results

species
character string specifying the name of the species. Only 'human', and 'mouse' are supported at present. [DEFAULT = human].

sample.names
character vector specifying sample names to use as prefix for files. If NULL, this will be derived from the file names. [DEFAULT = NULL]

no.model
logical, if TRUE then MACS2 shifting model will be bypassed.

extsize
numeric vector. If no.model = TRUE, this parameter is to extend reads in 5'->3' direction to fix-sized fragments. Ignored if no.model = FALSE or paired end reads. If used, MUST be same length as <treatment.files> and in same order. [DEFAULT = 0]

shift
numeric vector. If no.model = TRUE, this value is used to move cutting ends (5') then apply extsize from 5' to 3' direction to extend them to fragments. When this value is negative, ends will be moved toward 3'->5' direction, otherwise 5'->3' direction. [DEFAULT = 0]

no.lambda
logical, if TRUE then the background lambda will be used as local lambda.

call.summits
logical, if TRUE then MACS2 will reanalyze the shape of signal profile to deconvolve subpeaks within each peak. If used, the output subpeaks of a big peak region will have the same peak boundaries, but different scores and peak summit positions. [DEFAULT = TRUE]

qvalue
numeric value specifying the minimum FDR cutoff to call significant regions. [DEFAULT = 0.05]

verbosity
integer value specifying verbosity. If 0 then only critical messages displayed. If greater or equal to 3 then all messages shown.

threads
an integer value indicating the number of workers to be used. If NULL then one less than the maximum number of cores will be used. [DEFAULT = NULL].